+++ Sources/Flare/Classes/Models/SubscriptionInfo.swift
@@ -36,4 +36,10 @@ extension SubscriptionInfo: ISubscriptionInfo {
             try await self.underlyingSubscriptionInfo.subscriptionStatus
         }
     }
+
+    public var isEligibleForIntroOffer: SubscriptionEligibility {
+        get async {
+            await self.underlyingSubscriptionInfo.isEligibleForIntroOffer
+        }
+    }
 }