+++ Tests/SnapshotTests/ProductsViewSnapshotTests.swift
@@ -10,6 +10,7 @@ import Foundation
 
 // MARK: - ProductsViewSnapshotTests
 
+@MainActor
 @available(watchOS, unavailable)
 @available(iOS 13.0, macOS 10.15, tvOS 13.0, *)
 final class ProductsViewSnapshotTests: SnapshotTestCase {